pyproject.toml
[project]
name = "practice-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "databricks-langchain>=0.7.0",
    "grandalf>=0.8",
    "langchain[google-genai]>=0.3.27",
    "langgraph>=0.6.3",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
